Backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
email-validator
SQLAlchemy
psycopg2-binary
httpx

# Authentication
python-jose[cryptography]
bcrypt==4.0.1
passlib==1.7.4
python-multipart

# AI/LLM
langgraph
langchain
langchain-openai
typing_extensions
openai
qubipy
pyotp==2.9.0
qrcode[pil]==7.4.2
pandas
